requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
requests>=2.31.0
youtube_search>=2.1.2
googlesearch-python>=1.2.3
pydantic>=2.4.0

# Testing dependencies (optional)
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.25.0